requirements.txt
streamlit
openai
pypdf
sentence-transformers
python-dotenv
faiss-cpu
gtts
speechrecognition
pyaudio